Add epoch-enriched LeaderLatchListener
Add an epoch value to a new EpochLeaderLatchListener, which works
in the same way as LeaderLatchListener, but includes an epoch value
when leadership is gained.

The motivation for this change is to allow the leader to propagate
its epoch value to downstream services, which then allows those
downstream services from interacting with a stale leader based
on the epoch value being less than the maximum-seen epoch value.